requirements.txt
appnope==0.1.0
bleach==1.5.0
click==6.7
decorator==4.0.11
Django==1.10.3
django-extensions==1.7.5
djangorestframework==3.5.3
djangorestframework-gis==0.11
entrypoints==0.2.2
geocoder==1.20.0
gunicorn==19.6.0
html5lib==0.9999999
ipykernel==4.5.2
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.9.4
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.1
lxml==3.7.2
MarkupSafe==0.23
mistune==0.7.3
nbconvert==5.0.0
nbformat==4.2.0
notebook==4.3.1
numpy==1.12.0
pandas==0.19.2
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.6.0
pytz==2016.10
pyzmq==16.0.2
qtconsole==4.2.1
ratelim==0.1.6
requests==2.13.0
simplegeneric==0.8.1
six==1.10.0
terminado==0.6
testpath==0.3
tornado==4.4.2
traitlets==4.3.1
wcwidth==0.1.7
widgetsnbextension==1.2.6
